Improved project structure to attain separation of concern
