+++ DinnerDash/urls.py
@@ -21,5 +21,6 @@ from . import settings
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('items/', include('Items.urls'))
+    path('', include('Items.urls')),
+    path('', include('Order.urls'))
 ] + static(settings.MEDIA_URL, document_root = settings.MEDIA_ROOT)
