+++ Items/Models/category.py
@@ -1,0 +1,10 @@
+from django.db import models
+
+
+class Category(models.Model):
+    category_id = models.AutoField(primary_key=True)
+    name = models.CharField(max_length=150, unique=True)
+    description = models.TextField()
+
+    def __str__(self):
+        return self.name
