+++ Items/templates/items/login.html
@@ -14,7 +14,7 @@
             type="email"
             name="email"
             id=""
-            value="{{values.email}}"
+            value="{{email_value.email}}"
             class="form-control"
             placeholder="example@gmail.com"
             required
