+++ Items/templatetags/custom_filters.py
@@ -2,6 +2,7 @@ from django import template
 
 register = template.Library()
 
+
 @register.filter(name='multiply')
 def multiply(quantity, price):
     return quantity * price
