+++ Order/templates/order/cart.html
@@ -1,6 +1,6 @@
 {% extends 'items/base.html' %} {% block content %} 
 {% load cart %}
-{% load custom_filter %}
+{% load custom_filters %}
 <div class="container">
     <div class="border p-4 rounded m-4 col-lg-7 mx-auto">
         <div class="display-4 pb-2 text-success text-center font-weight-normal">Your Cart</div>
